move newsletter highlight cards into their own folder
